+++ app/javascript/components/game/map/MapDeployDisplay.tsx
@@ -27,16 +27,17 @@ export default function MapDeployDisplay({ scenario, scale, turn, player }: MapD
         const hexes = player === 1 ? (scenario.map.alliedSetupHexes as DeployHexesTurn)[turn] :
           (scenario.map.axisSetupHexes as DeployHexesTurn)[turn]
         const show = deployHex(hexes, x, y) ? scenario.map.baseTerrainColor : "#AAA"
+        const showStroke = deployHex(hexes, x, y) ? "#BBB" : "#777"
         hexLoader.push(
           <polygon key={`${x}-${y}-h`} points={hex.hexCoords}
-                   style={{ fill: show, stroke: "#777", strokeWidth: 2 }} />
+                   style={{ fill: show, stroke: showStroke, strokeWidth: 2 }} />
         )
         const vp = scenario.map.victoryAt(new Coordinate(x, y)) 
         if (vp) {
           const fill = `url(#nation-${vp === 1 ? scenario.alliedFactions[0] : scenario.axisFactions[0]}-16)`
           vpLoader.push(
             <circle transform={`scale(3) translate(${-hex.xOffset/1.5} ${-hex.yOffset/1.5})`} key={`${x}-${y}-vp`}
-                    cx={hex.xOffset} cy={hex.yOffset} r={16} style={{ fill, strokeWidth: 1, stroke: "black" }}/>
+                    cx={hex.xOffset} cy={hex.yOffset} r={16} style={{ fill, strokeWidth: 0 }}/>
           )
         }
       })